#2702 Removed "ParseTaxRates" from "Order" entity
