+++ src/Libraries/Nop.Core/Domain/Orders/Order.cs
@@ -1,8 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Globalization;
-using System.Linq;
 using Nop.Core.Domain.Common;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Discounts;
@@ -24,49 +21,6 @@ namespace Nop.Core.Domain.Orders
         private ICollection<OrderItem> _orderItems;
         private ICollection<Shipment> _shipments;
 
-        #region Utilities
-
-        /// <summary>
-        /// Parse tax rates
-        /// </summary>
-        /// <param name="taxRatesStr"></param>
-        /// <returns></returns>
-        protected virtual SortedDictionary<decimal, decimal> ParseTaxRates(string taxRatesStr)
-        {
-            var taxRatesDictionary = new SortedDictionary<decimal, decimal>();
-            if (string.IsNullOrEmpty(taxRatesStr))
-                return taxRatesDictionary;
-
-            var lines = taxRatesStr.Split(new [] { ';' }, StringSplitOptions.RemoveEmptyEntries);
-            foreach (var line in lines)
-            {
-                if (string.IsNullOrEmpty(line.Trim()))
-                    continue;
-
-                var taxes = line.Split(':');
-                if (taxes.Length == 2)
-                {
-                    try
-                    {
-                        var taxRate = decimal.Parse(taxes[0].Trim(), CultureInfo.InvariantCulture);
-                        var taxValue = decimal.Parse(taxes[1].Trim(), CultureInfo.InvariantCulture);
-                        taxRatesDictionary.Add(taxRate, taxValue);
-                    }
-                    catch (Exception exc)
-                    {
-                        Debug.WriteLine(exc.ToString());
-                    }
-                }
-            }
-
-            //add at least one tax rate (0%)
-            if (!taxRatesDictionary.Any())
-                taxRatesDictionary.Add(decimal.Zero, decimal.Zero);
-
-            return taxRatesDictionary;
-        }
-
-        #endregion
 
         #region Properties
 
@@ -490,7 +444,7 @@ namespace Nop.Core.Domain.Orders
         {
             get
             {
-                return ParseTaxRates(TaxRates);
+                return this.ParseTaxRates(TaxRates);
             }
         }
