+++ src/Libraries/Nop.Core/Domain/Orders/OrderExtensions.cs
@@ -1,0 +1,57 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.Linq;
+
+namespace Nop.Core.Domain.Orders
+{
+    /// <summary>
+    /// Customer extensions
+    /// </summary>
+    public static class OrderExtensions
+    {
+        /// <summary>
+        /// Parse tax rates
+        /// </summary>
+        /// <param name="order">Order</param>
+        /// <param name="taxRatesStr"></param>
+        /// <returns>Rates</returns>
+        public static SortedDictionary<decimal, decimal> ParseTaxRates(this Order order, string taxRatesStr)
+        {
+            var taxRatesDictionary = new SortedDictionary<decimal, decimal>();
+
+            if (string.IsNullOrEmpty(taxRatesStr))
+                return taxRatesDictionary;
+
+            var lines = taxRatesStr.Split(new[] { ';' }, StringSplitOptions.RemoveEmptyEntries);
+            foreach (var line in lines)
+            {
+                if (string.IsNullOrEmpty(line.Trim()))
+                    continue;
+
+                var taxes = line.Split(':');
+                if (taxes.Length == 2)
+                {
+                    try
+                    {
+                        var taxRate = decimal.Parse(taxes[0].Trim(), CultureInfo.InvariantCulture);
+                        var taxValue = decimal.Parse(taxes[1].Trim(), CultureInfo.InvariantCulture);
+                        taxRatesDictionary.Add(taxRate, taxValue);
+                    }
+                    catch (Exception exc)
+                    {
+                        Debug.WriteLine(exc.ToString());
+                    }
+                }
+            }
+
+            //add at least one tax rate (0%)
+            if (!taxRatesDictionary.Any())
+                taxRatesDictionary.Add(decimal.Zero, decimal.Zero);
+
+            return taxRatesDictionary;
+        }
+
+    }
+}
